fix paths, move Header, Dropdown and Footer to Global
