+++ docs/components/Global/Header/Tagline.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Paragraph } from '../../Typography';
+
+export function Tagline() {
+  return (
+    <Paragraph textAlign="center">
+      a light, modular and extensible ReactJS UI primitives library
+    </Paragraph>
+  );
+}
