--- docs/components/Header/Tagline.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Paragraph } from '../Typography';
-
-export function Tagline() {
-  return <Paragraph textAlign="center">a light, modular and extensible ReactJS UI primitives library</Paragraph>
-}
